first crash with scene delegate [SceneConfiguration] Info.plist configuration "Default Configuration" for UIWindowSceneSessionRoleApplication contained UISceneDelegateClassName key, but could not load class with name "VseBankiRu.SceneDelegate".

+++ VseBankiRu/Modules/CreditsListFilter/Model/FilterItemModel.swift
@@ -11,7 +11,7 @@ import Foundation
 struct FilterItemModel {
     var bankName: String? //банк
     var goal: String? //цель
-    var time: Int? //срок
+    var time: Double? //срок
     var maxValue: Int //максимальная сумма
     var minValue: Int //минимальная сумма
     var value: Int //сумма
